make a info file in consensus
