+++ consensus/state.go
@@ -1,8 +1,6 @@
 package consensus
 
 import (
-	"errors"
-	"fmt"
 	"math/big"
 	"sync"
 )
@@ -94,130 +92,6 @@ func CreateGenesisState() (s *State) {
 	s.unspentOutputs[genesisBlock.SubsidyID()] = genesisSubsidyOutput
 
 	return
-}
-
-// BlockAtHeight() returns the block from the current history at the
-// input height.
-func (s *State) blockAtHeight(height BlockHeight) (b Block, err error) {
-	if bn, ok := s.blockMap[s.currentPath[height]]; ok {
-		b = bn.Block
-		return
-	}
-	err = fmt.Errorf("no block at height %v found.", height)
-	return
-}
-
-// currentBlockNode returns the node of the most recent block in the
-// longest fork.
-func (s *State) currentBlockNode() *BlockNode {
-	return s.blockMap[s.currentBlockID]
-}
-
-// CurrentBlockWeight() returns the weight of the current block in the
-// heaviest fork.
-func (s *State) currentBlockWeight() BlockWeight {
-	return s.currentBlockNode().Target.Inverse()
-}
-
-// depth returns the depth of the current block of the state.
-func (s *State) depth() Target {
-	return s.currentBlockNode().Depth
-}
-
-// height returns the current height of the state.
-func (s *State) height() BlockHeight {
-	return s.blockMap[s.currentBlockID].Height
-}
-
-// State.Output returns the Output associated with the id provided for input,
-// but only if the output is a part of the utxo set.
-func (s *State) output(id OutputID) (output Output, exists bool) {
-	output, exists = s.unspentOutputs[id]
-	return
-}
-
-// Block returns the block associated with the given id.
-func (s *State) Block(id BlockID) (b Block, exists bool) {
-	node, exists := s.blockMap[id]
-	if !exists {
-		return
-	}
-	b = node.Block
-	return
-}
-
-// BlockAtHeight returns the block in the current fork found at `height`.
-func (s *State) BlockAtHeight(height BlockHeight) (b Block, err error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	bn, exists := s.blockMap[s.currentPath[height]]
-	if !exists {
-		err = errors.New("no block found")
-		return
-	}
-	b = bn.Block
-	return
-}
-
-func (s *State) Contract(id ContractID) (fc FileContract, err error) {
-	fc, exists := s.openContracts[id]
-	if !exists {
-		err = errors.New("no contract found")
-		return
-	}
-
-	return
-}
-
-// CurrentBlock returns the highest block on the tallest fork.
-func (s *State) CurrentBlock() Block {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.currentBlockNode().Block
-}
-
-// CurrentTarget returns the target of the next block that needs to be
-// submitted to the state.
-func (s *State) CurrentTarget() Target {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.currentBlockNode().Target
-}
-
-// EarliestLegalTimestamp returns the earliest legal timestamp of the next
-// block - earlier timestamps will render the block invalid.
-func (s *State) EarliestTimestamp() Timestamp {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.currentBlockNode().earliestChildTimestamp()
-}
-
-// State.Height() returns the height of the longest fork.
-func (s *State) Height() BlockHeight {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.height()
-}
-
-// HeightOfBlock returns the height of the block with id `bid`.
-func (s *State) HeightOfBlock(bid BlockID) (height BlockHeight, err error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	bn, exists := s.blockMap[bid]
-	if !exists {
-		err = errors.New("block not found")
-		return
-	}
-	height = bn.Height
-	return
-}
-
-func (s *State) Output(id OutputID) (output Output, exists bool) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	return s.output(id)
 }
 
 func (s *State) RLock() {
